refactor: 프로그레스바 파일 분리하여 컴포넌트화
