+++ src/pages/detail/DetailPage.jsx
@@ -1,9 +1,10 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import { useParams } from 'react-router';
 import { useMutation, useQuery, useQueryClient } from 'react-query';
 import { getNewsDetail, postUpdateLike } from '../../api/newsDetail';
-import { PostHead, ProgressBar, PostContainer, PostHashtag } from './DetailPageStyle';
+import { PostHead, PostContainer, PostHashtag } from './DetailPageStyle';
 import LoadingPage from '../../components/loadingPage/LoadingPage';
+import ProgressBar from '../../components/progressBar/ProgressBar';
 import DetailPageLike from '../../components/detailPageLike/DetailPageLike';
 import DetailPageSubscribe from '../../components/detailPageSubscribe/DetailPageSubscribe';
 import HomeBanner from '../../components/homeBanner/HomeBanner';
@@ -25,29 +26,6 @@ const DetailPage = () => {
         mutation.mutate(id);
     }
 
-    const [scrollPercentage, setScrollPercentage] = useState(0);
-    const [isScrolling, setIsScrolling] = useState(false);
-    useEffect(() => {
-        const scrollListener = () => {
-            const element = document.documentElement;
-            const totalHeight = element.scrollHeight - element.clientHeight;
-            const scrollPosition = window.scrollY;
-            const percentageScrolled = (scrollPosition / totalHeight) * 100;
-
-            setScrollPercentage(percentageScrolled);
-
-            if (window.scrollY > 0) {
-                setIsScrolling(true);
-            } else {
-                setIsScrolling(false);
-            }
-        };
-
-        window.addEventListener('scroll', scrollListener);
-
-        return () => window.removeEventListener('scroll', scrollListener);
-    }, []);
-
     if (isLoading) {
         return <LoadingPage />;
     }
@@ -60,12 +38,7 @@ const DetailPage = () => {
         <section className="post is-sticky">
             <div className="post-scrollwrap">
                 <PostHead className="post-head">
-                    {isScrolling && (
-                        <ProgressBar>
-                            <div className="post-head-progress" role="progressbar" aria-valuenow={scrollPercentage} style={{ width: `${scrollPercentage}%` }}></div>
-                            <span className='progress-title'>{post.title}</span>
-                        </ProgressBar>
-                    )}
+                    <ProgressBar title={post.title} />
                     <a className="post-head-runninghead" href="/tag/domestic-issue">{post.category}</a>
                     <h2 className="post-head-headline">{post.title}</h2>
                     <time className="post-head-date">{post.date}</time>
